rsync 3.1.2
Update to rsync 3.1.2

As of 3.1.2, upstream provides "a tar file of the 'patches' directory
now released in a separate file" (https://rsync.samba.org/), so this
commit restructures the formula to use the new patches tar file instead
of the MacPorts and DomT4 urls for each patch separately.

To specify which patch files from the tar file are required and the
order in which they should be applied, the new "apply" method must be
used.

Depends on Homebrew/homebrew#47614

Closes #543.

Signed-off-by: Mike McQuaid <mike@mikemcquaid.com>