Revert "update awk formula"
This reverts commit 4dfe8686d97b8942d197d4ff4b5957471ea2aaff.

I'm not happy with this as-is. In the original discussion we advised OP to host
this formula in their own tap, but we've ended up with it replacing the dupes
formula with minimal discussion.

Stylistically the formula is all over the shop. It hasn't been through CI in
~6 months or more so it cannot possibly be known if it'll actually build today, or passes
audit/etc. The secure URLs were downgraded without any justification. There's merit
to discussing whether we even qualify `head` as a valid upstream source we're happy
with.

I should have closed this at the time it was submitted with similar comments, but
I let it sit, so I accept that's my fault for not being clear that I wasn't
thrilled as-is.